website/backend/pyproject.toml
[project]
name = "backend"
version = "0"
requires-python = ">=3.13"
dependencies = []

[tool.uv]
default-groups = ["skeleton", "game", "db"]  # excludes: dev

[dependency-groups]
db = [
    "sqlmodel>=0.0.24",
    "google-cloud-storage>=3.2.0",
    "xxhash>=3.5.0",
    "gcloud-aio-storage>=9.5.0",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.21.0",
    "cloud-sql-python-connector[asyncpg]>=1.18.3",
    "greenlet>=3.2.3",
]
dev = [
    "pytest>=8.4.0",
    "ruff>=0.11.13",
]
game = [
    "ale-py>=0.11.1",
    "distinctipy>=1.3.4",
    "gymnasium>=1.1.1",
    "networkx>=3.5",
    "opencv-python>=4.11.0.86",
    "pygame>=2.6.1",
]
skeleton = [
    "fastapi[standard]>=0.115.12",
    "python-dotenv>=1.1.0",
    "websockets>=15.0.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "N",   # pep8-naming
    "I",   # isort
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
    "SIM", # flake8-simplify
    "C90", # mccabe cyclomatic complexity
]
ignore = [
    "S101",  # assert
]
